fix: save recurring payments as previous payment method
There was a bug related to the onSuccess callback on the reserveMutation
in PurchaseConfirmationScreen that broke saving of recurring payment.

It seems the root of the issue was a "stale closure" problem, and works
as follows:

1. A closure was provided as useReserveOfferMutation's onSuccess, that
   had a callback to onPaymentCompleted.
2. onPaymentCompleted is a useCallback, with a dependency array updating
   the function every time recurringPaymentId changes.
3. The assumption was that once recurringPaymentId changes, the
   onSuccess closure would update as well. But this is not the case,
   since onSuccess itself doesn't have a dependency array, and becomes
   stale if any state other that the argument (reserveOfferResponse)
   changes.

I believe this is why the onSuccess, onError and onSettled callbacks are
deprecated and removed in TanStack Query v5.
https://tanstack.com/query/v4/docs/framework/react/reference/useQuery