feat: add publicCode to global message contexts
Adds publicCode as string to departure details, trip details and trip
results, so that we can add global messages for specific line numbers.

Also adds publicCodeAsNumber to departure details, so line numbers in a
range can be targeted, which can be useful for cases like 100-999 for
regional buses. This is harder to do with the rule engine when there is
more than one public code.